GOCA - split file goca.go between client.go and controller.go
